+++ packages/graphiql/cypress/support/index.js
@@ -12,7 +12,7 @@
 // You can read more here:
 // https://on.cypress.io/configuration
 // ***********************************************************
-// / <reference types="cypress" />
+/// <reference types="cypress" />
 
 // import '@cypress/code-coverage/support';