+++ packages/monaco-graphql/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 1.1.6
+
+### Patch Changes
+
+- [#2900](https://github.com/graphql/graphiql/pull/2900)
+  [`8989ffce`](https://github.com/graphql/graphiql/commit/8989ffce7d6beca874e70f5a1ff066102580173a)
+  Thanks [@acao](https://github.com/acao)! - use decorators-legacy @babel/parser
+  plugin so that all styles of decorator usage are supported
+
 ## 1.1.5
 
 ### Patch Changes